refactor(workflow): extract planning state codecs
